+++ examples/attribute-policies/client.py
@@ -45,9 +45,10 @@ assert not gh.check(Inquiry(action='off',
 
 
 # administration panel requests
-assert not gh.check(Inquiry(action='fork', resource='repos/DataDog/kafka-kit', subject={'name': 'Jim', 'stars': 1}))
+assert not gh.check(Inquiry(action='fork', resource='repos/datadog/kafka-kit', subject={'name': 'Jim', 'stars': 1}))
 assert gh.check(Inquiry(action='fork', resource='repos/DataDog/kafka-kit', subject={'name': 'Ho', 'stars': 80}))
 assert gh.check(Inquiry(action='fork', resource='repos/DataDog/consul', subject={'name': 'Ho', 'stars': 998}))
+assert gh.check(Inquiry(action='fork', resource='repos/datadog/consul', subject={'name': 'Ho', 'stars': 998}))
 assert not gh.check(Inquiry(action='fork', resource='repos/DataDog/consul', subject={'name': 'Ho', 'stars': 999}))